Refactor: Template Data stored in indexedDB
Previously using local storage (JSON.stringify) synchronous storage.
The idea was to only hold a single workbook of data at a time,
regardless of which template was being edited by the user.
Using indexedDB we have less constraints in terms of storage and now
opting for data to belong to a specific template.
Asynchronous nature of indexedDB means it makes sense to use react-query
for data fetching so new templateDataQueries have been introduced